Create Workflow of 5G and UE
We already installed Open5gs and UERANSIM in last sprint. In this sprint, we started configuring ngap address and gtpu address in amf.yaml and upf.yaml files. Here we also make a decision on whether to use separate servers or VMs for Open5gs and UERANSIM or remain on one server.